backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-genai
python-multipart
aiohttp
requests
reportlab
pydantic
python-dotenv
pefile
yara-python